requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
mysql-connector-python==9.2.0
propcache==0.2.1
py-mon==2.0.5
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
watchdog==6.0.0
yarl==1.18.3
